requirements.txt
playwright
behave
python-dotenv
PyYAML
